layers: Move DynamicState CoreCheck to own file
